rename package with snack bar
