+++ app/src/main/java/com/sanaa/tudee_assistant/presentation/screen/add_edit_screen/TaskUiState.kt
@@ -1,6 +1,7 @@
-package com.sanaa.tudee_assistant.presentation.screen.addEditScreen
+package com.sanaa.tudee_assistant.presentation.screen.add_edit_screen
 
 import com.sanaa.tudee_assistant.domain.model.Category
+import com.sanaa.tudee_assistant.presentation.model.CategoryState
 import com.sanaa.tudee_assistant.presentation.model.TaskUiPriority
 import com.sanaa.tudee_assistant.presentation.model.TaskUiStatus
 import com.sanaa.tudee_assistant.presentation.state.TaskUiModel
@@ -15,11 +16,13 @@ data class AddTaskUiState(
         description = "",
         dueDate = null,
         categoryId = -1,
+
         categoryImagePath = "",
         priority = TaskUiPriority.LOW,
         status = TaskUiStatus.TODO,
         createdAt = Clock.System.now().toLocalDateTime(TimeZone.currentSystemDefault())
     ),
+    val categories: List<CategoryState> = emptyList(),
     val selectedCategory: Category? = null,
     val isButtonEnabled: Boolean = false,
     val isOperationSuccessful: Boolean = false,
